Got long sends mostly working. There's one bug left though: In the pingpong app, the client doesn't use the server's return buffer for its reply, even though it receives it correctly.
Also:
* Use memset() instead of bzero(), as the latter is deprecated
* Use size_t for buffer sizes
* When allocating buffers, return NULL if size is 0

+++ clients/client.c
@@ -13,7 +13,7 @@ int main(void) {
 	void *msg_array[2];
 	msg_array[0] = ripc_buf_alloc(sizeof(i));
 	msg_array[1] = ripc_buf_alloc(strlen(msg));
-	int length_array[2];
+	size_t length_array[2];
 	for (i = 0; true; ++i) {
 		*(int *)msg_array[0] = i;
 		length_array[0] = sizeof(i);
